// types/dengue.ts
export interface DengueCase {
  tipoNotificacao: string
  agravo: string
  dataNotificacao: string
  ufNotificacao: string
  municipioNotificacao: string
  unidadeSaude: string
  dataPrimeirosSintomas: string
  anoNascimento: string
  sexo: string
  gestante: string
  raca: string
  ufResidencia: string
  municipioResidencia: string
  paisResidencia: string
  dataInvestigacao: string
  febre: string
  mialgia: string
  cefaleia: string
  exantema: string
  vomito: string
  nausea: string
  dorCostas: string
  conjuntivite: string
  artrite: string
  artralgia: string
  petequias: string
  leucopenia: string
  testeLaco: string
  dorRetroOrbital: string
  diabetes: string
  doencaHematologica: string
  hepatopatia: string
  doencaRenal: string
  hipertensao: string
  ulceraPeptica: string
  doencaAutoimune: string
  hospitalizacao: string
  dataInternacao: string
  dataObito: string
  [key: string]: any
}